refactor: entity 이름과 패키지 변경 (#166)
